gdqcli: Add flags to show category and platform
This adds two flags to respective show the category and the platform for
the event. They're hidden by default because they're not as interesting
to most people, and category is extremely messy which causes text to
overflow on even wide terminals.

The new writer is a messy hack. It only really works because we only
have 2 hidden fields, if we had more we'd need to take a different
approach to this.