schedule: Check error after every Find call
Before we were assuming that the chain of Find.Find.FindAll would always
return successfully. Now that through the CLI someone can put in an
invalid schedule ID, this breaks and we get a runtime panic. Instead,
after every Find call check the Error on soup.Root and bail out if it's
not nil.