Rewrite the library to use the tracker API
The first version of the library used the /schedule endpoint to scrape
the HTML. It turns out this isn't necessary, as schedule itself is built
using the /tracker/api endpoints. Using /search and /hosts we can get
proper JSON instead.

This rewrites the library to use the Tracker API instead.